일정 유저 연관관계 수정

+++ src/main/java/com/jyp/justplan/domain/plan/application/PlanService.java
@@ -14,6 +14,10 @@ import com.jyp.justplan.domain.plan.dto.request.PlanIdRequest;
 import com.jyp.justplan.domain.plan.dto.request.PlanCreateRequest;
 import com.jyp.justplan.domain.plan.dto.request.PlanUpdateRequest;
 import com.jyp.justplan.domain.plan.dto.response.*;
+import com.jyp.justplan.domain.plan.exception.PlanValidationException;
+import com.jyp.justplan.domain.user.UserDetailsImpl;
+import com.jyp.justplan.domain.user.application.UserService;
+import com.jyp.justplan.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -34,12 +38,16 @@ public class PlanService {
     private final PlanTagService planTagService;
     private final BudgetService budgetService;
     private final ExpenseService expenseService;
+    private final UserPlanService userPlanService;
+    private final UserService userService;
 
     /* Plan을 통한 PlanResponse 반환 (origin Plan 정보 포함) */
     private PlanDetailResponse getPlanDetailResponse(Plan plan) {
         // 일정에 해당하는 태그 조회
         List<PlanTag> tags = planTagService.findTagsByPlan(plan);
         List<String> tagNames = getTagNames(tags);
+
+        List<UserPlanResponse> users = userPlanService.findUserPlanResponsesByPlan(plan);
 
         CityResponse cityResponse = new CityResponse(plan.getRegion());
 
@@ -54,19 +62,21 @@ public class PlanService {
             List<String> originTagNames = originTags.stream()
                     .map(tag -> tag.getTag().getName())
                     .collect(Collectors.toList());
+            List<UserPlanResponse> originUsers = userPlanService.findUserPlanResponsesByPlan(plan.getOriginPlan());
 
-            PlanResponse originPlan = PlanResponse.toDto(plan.getOriginPlan(), originTagNames);
-            return PlanDetailResponse.toDto(plan, tagNames, cityResponse, originPlan, budgetResponse, expenseResponse);
+            PlanResponse originPlan = PlanResponse.toDto(plan.getOriginPlan(), originUsers, originTagNames);
+            return PlanDetailResponse.toDto(plan, users, tagNames, cityResponse, originPlan, budgetResponse, expenseResponse);
         } else {
-            return PlanDetailResponse.toDto(plan, tagNames, cityResponse, budgetResponse, expenseResponse);
+            return PlanDetailResponse.toDto(plan, users, tagNames, cityResponse, budgetResponse, expenseResponse);
         }
     }
 
     private PlanResponse getPlanResponse(Plan plan) {
         List<PlanTag> tags = planTagService.findTagsByPlan(plan);
         List<String> tagNames = getTagNames(tags);
+        List<UserPlanResponse> users = userPlanService.findUserPlanResponsesByPlan(plan);
 
-        return PlanResponse.toDto(plan, tagNames);
+        return PlanResponse.toDto(plan, users, tagNames);
     }
 
     /* 전체 플랜 조회 */
@@ -98,11 +108,14 @@ public class PlanService {
 
     /* 플랜 생성 */
     @Transactional
-    public PlanDetailResponse savePlan(PlanCreateRequest request) {
+    public PlanDetailResponse savePlan(PlanCreateRequest request, String userEmail) {
         Plan plan = planRepository.save(request.toEntity());
+        User user = userService.findByEmail(userEmail);
 
         City city = cityRepository.getById(request.getRegionId());
         plan.setRegion(city);
+
+        userPlanService.saveUserPlan(user, plan, true);
 
         budgetService.createBudget(plan);
         expenseService.createExpense(plan);
@@ -114,8 +127,10 @@ public class PlanService {
 
     /* 플랜 복제 (가져오기) */
     @Transactional
-    public PlanDetailResponse copyPlan(PlanIdRequest request) {
+    public PlanDetailResponse copyPlan(PlanIdRequest request, String userEmail) {
         Plan origin_plan = planRepository.getById(request.getOriginPlanId());
+        User user = userService.findByEmail(userEmail);
+
         String title = origin_plan.getTitle() + " 복사본";
 
         Plan new_plan =  planRepository.save(new Plan(
@@ -126,6 +141,8 @@ public class PlanService {
 
         new_plan.setRegion(origin_plan.getRegion());
         new_plan.setOriginPlan(origin_plan);
+
+        userPlanService.saveUserPlan(user, new_plan, true);
 
         List<PlanTag> origin_tags = planTagService.findTagsByPlan(origin_plan);
         planTagService.savePlanTag(new_plan, getTagNames(origin_tags));
@@ -142,8 +159,12 @@ public class PlanService {
     /* 플랜 수정 (제목, 여행 일자, 태그) */
     /* + 예산, 지출, 공개 여부, 지출 여부 수정 */
     @Transactional
-    public PlanDetailResponse updatePlan(PlanUpdateRequest request) {
+    public PlanDetailResponse updatePlan(PlanUpdateRequest request, String userEmail) {
         Plan plan = planRepository.getById(request.getPlanId());
+        User user = userService.findByEmail(userEmail);
+
+        validateUserOfPlan(plan, user);
+
         List<String> tags = request.getTags();
 
         // 태그 수정
@@ -169,10 +190,25 @@ public class PlanService {
 
     /* 플랜 삭제 */
     @Transactional
-    public void deletePlan(Long planId) {
+    public void deletePlan(Long planId, String userEmail) {
         Plan plan = planRepository.getById(planId);
+        User user = userService.findByEmail(userEmail);
+
+        validateOwnerOfPlan(plan, user);
 
 //        planTagService.deletePlanTag(plan);
         planRepository.delete(plan);
+    }
+
+    public void validateOwnerOfPlan (Plan plan, User user) {
+        if (!userPlanService.findOwnerByPlan(plan).equals(user)) {
+            throw new PlanValidationException("해당 일정의 소유자가 아닙니다.");
+        }
+    }
+
+    public void validateUserOfPlan (Plan plan, User user) {
+        if (!userPlanService.findUsersByPlan(plan).contains(user)) {
+            throw new PlanValidationException("해당 일정의 참여자가 아닙니다.");
+        }
     }
 }
